changed sleep/power button priority
I have a suspicion that in the absence of a hardware sleep button, these two settings buttons refer to the same memory location and at the same time the sleep button had to act on a single button. let the action assigned to the power button be a higher priority in this case.